Completely refactored again to facilitate new simple batch jobs
